backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
SQLAlchemy==2.0.23
Flask-Admin==1.6.1
pycryptodomex==3.19.0
Flask-SQLAlchemy==3.1.1
Pillow==10.1.0
-e .
